Move some modules, create providers dir Set min version for duckduckgo Make duckduckgo search async Remove get_lastet_version
