+++ etc/unittest/__main__.py
@@ -3,5 +3,7 @@ from .asyncio import *
 from .backend import *
 from .main import *
 from .model import *
+from .client import *
+from .include import *
 
 unittest.main()
