+++ etc/unittest/backend.py
@@ -1,9 +1,11 @@
 import unittest
+# import asyncio
 from unittest.mock import MagicMock
 from .mocks import ProviderMock
 import g4f
 try:
     from g4f.gui.server.backend import Backend_Api, get_error_message
+    # from g4f.gui.server.internet import search
     has_requirements = True
 except:
     has_requirements = False
@@ -16,10 +18,10 @@ class TestBackendApi(unittest.TestCase):
         self.app = MagicMock()
         self.api = Backend_Api(self.app)
 
-    def test_version(self):
-        response = self.api.get_version()
-        self.assertIn("version", response)
-        self.assertIn("latest_version", response)
+    # def test_version(self):
+    #     response = self.api.get_version()
+    #     self.assertIn("version", response)
+    #     self.assertIn("latest_version", response)
         
     def test_get_models(self):
         response = self.api.get_models()
@@ -30,6 +32,10 @@ class TestBackendApi(unittest.TestCase):
         response = self.api.get_providers()
         self.assertIsInstance(response, list)
         self.assertTrue(len(response) > 0)
+        
+    # def test_search(self):
+    #     result = asyncio.run(search("Hello"))
+    #     self.assertEqual(5, len(result))
         
 class TestUtilityFunctions(unittest.TestCase):
