+++ g4f/models.py
@@ -1,5 +1,7 @@
 from __future__  import annotations
+
 from dataclasses import dataclass
+
 from .Provider   import RetryProvider, ProviderType
 from .Provider   import (
     Chatgpt4Online,
