+++ g4f/providers/retry_provider.py
@@ -2,8 +2,9 @@ from __future__ import annotations
 
 import asyncio
 import random
+
 from ..typing import CreateResult, Messages
-from ..base_provider import BaseRetryProvider
+from .types import BaseRetryProvider
 from .. import debug
 from ..errors import RetryProviderError, RetryNoProviderError
