+++ requirements.txt
@@ -1,11 +1,11 @@
 requests
 pycryptodome
-curl_cffi>=0.5.10
+curl_cffi>=0.6.0b9
 aiohttp
 certifi
 browser_cookie3
 PyExecJS
-duckduckgo-search
+duckduckgo-search>=4.4.3
 nest_asyncio
 werkzeug
 loguru
